[hyperion] Send instance-scoped commands one instance at a time
Hyperion.ng's componentstate command rejects an instance array with more than one entry and also requires instance to be an array, so instance-scoped commands are now sent once per targeted instance using a single-element array. Instance targeting is lifted into the base HyperionCommand so brightness, color, effect and clear commands also apply to all configured instances instead of only instance 0.
