+++ bundles/org.openhab.binding.hyperion/src/main/java/org/openhab/binding/hyperion/internal/protocol/HyperionCommand.java
@@ -12,6 +12,9 @@
  */
 package org.openhab.binding.hyperion.internal.protocol;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import com.google.gson.annotations.SerializedName;
 
 /**
@@ -24,6 +27,9 @@ public abstract class HyperionCommand {
 
     @SerializedName("command")
     private String command;
+    // Hyperion.ng targets instances via an array of instance indices
+    @SerializedName("instance")
+    private List<Integer> instance;
 
     public HyperionCommand(String command) {
         setCommand(command);
@@ -35,5 +41,21 @@ public abstract class HyperionCommand {
 
     public void setCommand(String command) {
         this.command = command;
+    }
+
+    public void setInstance(List<Integer> instanceList) {
+        if (instanceList == null || instanceList.isEmpty()) {
+            this.instance = null;
+            return;
+        }
+        List<Integer> copy = new ArrayList<>(instanceList.size());
+        for (Integer v : instanceList) {
+            if (v == null || v < 0) {
+                this.instance = null;
+                return;
+            }
+            copy.add(v);
+        }
+        this.instance = copy;
     }
 }
